+++ x/btcstaking/types/validate_parsed_message_test.go
@@ -88,7 +88,6 @@ func generateUnbondingInfo(
 	unbondingValue int64,
 	p *types.Params,
 ) *unbondingInfo {
-
 	covPKs, err := bbn.NewBTCPKsFromBIP340PKs(p.CovenantPks)
 	require.NoError(t, err)
 
@@ -222,7 +221,7 @@ func createMsgDelegationForParams(
 		SlashingTx:                    testStakingInfo.SlashingTx,
 		DelegatorSlashingSig:          delegatorSig,
 		UnbondingTx:                   unbondingInfo.serializedUnbondingTx,
-		UnbondingTime:                 uint32(unbondingTime),
+		UnbondingTime:                 unbondingTime,
 		UnbondingValue:                unbondingValue,
 		UnbondingSlashingTx:           unbondingInfo.unbondingSlashingTx,
 		DelegatorUnbondingSlashingSig: unbondingInfo.unbondingSlashinSig,
